H2_Database and Exception Handling

+++ src/main/java/com/bridgelabzMySqlDatabase/MySqlDatabase/Repository/MessageRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MessageRepo  extends JpaRepository<Message,Integer> {
+public interface MessageRepo  extends JpaRepository<Message,Long> {
 }
